% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adductForm.r
\name{adductForm}
\alias{adductForm}
\title{Determine the adduct forms}
\usage{
adductForm(mass, mz, ion_mode = "positive", mass_accuracy = 10)
}
\arguments{
\item{mass}{A numeric vector of monoisotopic masses.}

\item{mz}{A numeric vector of m/z values.}

\item{ion_mode}{A character string indicating the ion mode. Default is "positive".}

\item{mass_accuracy}{A numeric value indicating the mass accuracy in ppm. Default is 10.}
}
\value{
A data frame with the mass, mz, adduct, and ppm error.
}
\description{
This function determines the adduct forms given the provided monoisotopic masses and m/z.
}
