% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass2mz.R
\name{mass2mz}
\alias{mass2mz}
\title{Calculate mass-to-charge ratio from a monoisotopic mass}
\usage{
mass2mz(x, adduct = "M+H")
}
\arguments{
\item{x}{A numeric vector of monoisotopic masses.}

\item{adduct}{A character string indicating the adduct. Default is "M+H". If multiple adducts are provided, they should be provided as a vector, like c("M+H", "M+Na").}
}
\value{
A data frame with the mass, adduct, and mz.
}
\description{
\code{mass2mz} calculates the mass-to-charge ratio from a monoisotopic mass in the wide format.
}
