% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass2mz_df.R
\name{mass2mz_df}
\alias{mass2mz_df}
\title{Calculate mass-to-charge ratio from a formula}
\usage{
mass2mz_df(mass, adduct = "M+H")
}
\arguments{
\item{mass}{A numeric vector of monoisotopic masses.}

\item{adduct}{A character string indicating the adduct. Default is "M+H".}
}
\value{
A data frame with the mass, adduct, and mz.
}
\description{
\code{mass2mz_df} calculates the mass-to-charge ratio from a formula.
}
