% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz2mass_df.R
\name{mz2mass_df}
\alias{mz2mass_df}
\title{Calculate monoisotopic mass from a mass-to-charge ratio}
\usage{
mz2mass_df(x, adduct = "M+H")
}
\arguments{
\item{x}{A numeric vector of mass-to-charge ratios.}

\item{adduct}{A character string indicating the adduct. Default is "M+H". If multiple adducts are provided, they should be provided as a vector, like c("M+H", "M+Na").}
}
\value{
A data frame with the mz, adduct, and adduct_mass.
}
\description{
\code{mz2mass_df} calculates the monoisotopic mass from a mass-to-charge ratio in the long format.
}
